requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.49
botocore==1.31.70
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
django-cors-headers==4.2.0
django-filter==23.2
django-guardian==2.4.0
django-mysql==4.11.0
django-ses==3.5.0
django-slack==5.19.0
django-templated-mail==1.1.1
Django==4.2.7
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
djoser==2.2.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
jmespath==1.0.1
Markdown==3.5
multidict==6.0.4
mysqlclient==2.2.0
oauthlib==3.2.2
openai==0.28.1
pycparser==2.21
Pygments==2.16.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests-oauthlib==1.3.1
requests==2.31.0
rollbar==0.16.3
s3transfer==0.6.1
sentry-sdk==1.29.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==5.5.0
tqdm==4.65.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
uuid==1.30
uWSGI==2.0.22
yarl==1.9.2
django-oauth-toolkit==2.3.0
oauth2client==4.1.3
google-auth==2.23.3
